Modified park names
