+++ src/cms/api/protected-area/controllers/protected-area.js
@@ -1,7 +1,6 @@
 "use strict";
 const { sanitizeEntity } = require("strapi-utils");
 const customStatus = require("../custom/protected-area-status");
-const customName = require("../custom/protected-area-names");
 
 /**
  * Read the documentation (https://strapi.io/documentation/developer-docs/latest/development/backend-customization.html#core-controllers)
@@ -15,10 +14,33 @@ module.exports = {
     return sanitizeEntity(entity, { model: strapi.models["protected-area"] });
   },
   async names(ctx) {
-    return customName.getProtectedAreaNames(ctx);
+    // custom route for basic park details with park names
+    const parkNamesData = await strapi.services["park-name"].names();
+    const entities = await strapi.services["protected-area"].names(ctx);
+    return entities.map((entity) => {
+      const { id, orcs, type, typeCode, protectedAreaName } = sanitizeEntity(
+        entity,
+        {
+          model: strapi.models["protected-area"],
+        }
+      );
+      const parkNamesFilter = parkNamesData.filter((x) => x.orcs == orcs);
+
+      const parkNames =
+        parkNamesFilter.length !== 0 ? parkNamesFilter[0].parkNames : [];
+
+      return { id, orcs, type, typeCode, protectedAreaName, parkNames };
+    });
   },
-  async items(ctx) {
-    return customName.getProtectedAreaItems(ctx);
+  async items() {
+    // custom route for light weight park details used in client app
+    const entities = await strapi.services["protected-area"].items();
+    return entities.map((entity) => {
+      const { id, orcs, protectedAreaName } = sanitizeEntity(entity, {
+        model: strapi.models["protected-area"],
+      });
+      return { id, orcs, protectedAreaName };
+    });
   },
   async status(ctx) {
     return customStatus.getProtectedAreaStatus(ctx);
